Rename files
Renamed two files which were named for drawing custom areas but relate more widely to location search:
* `adHocPolygon.js` -> `locationSearchUI.js`
* `search_location_polygon_editing.html` -> `search_location_with_modes.html`
