+++ opentreemap/treemap/templates/treemap/map.html
@@ -10,7 +10,7 @@
 {% block search %}
 <div class="search-block-wrapper">
   {% include "treemap/partials/search_species.html" %}
-  {% include "treemap/partials/search_location_polygon_editing.html" %}
+  {% include "treemap/partials/search_location_with_modes.html" %}
 </div>
 {% endblock search %}
